investigation type prediction and patch
wrote script to 1) train the logreg classifier, for investigation type
only so far (I got 91.5% accuracy on a test group of 1M datasets on my
local machine), and 2) to iterate through every row in the database and
infer investigation type manually, if possible, or to predict it if
not.
-TODO right now just saves the new columns
‘metaseek_investigation_type’ and ‘P_investigation_type’ to a csv. Need
to insert the columns and do the database migration still.