Fix CSV downloads
Make the CSV files on the backend, send the finished file for simple
download at:

/discovery/<int:id>/download
/discovery/<int:id>/ids

TODO - those responses can be VERY slow to generate for large
discoveries. Maybe put these in a background job and email users when
the result is finished. Or at minimum a spinner in the UI.